Rename postingiterator* -> posting_iterator*.
