+++ searchlib/src/vespa/searchlib/memoryindex/CMakeLists.txt
@@ -10,7 +10,7 @@ vespa_add_library(searchlib_memoryindex OBJECT
     field_inverter.cpp
     memory_index.cpp
     ordered_field_index_inserter.cpp
-    postingiterator.cpp
+    posting_iterator.cpp
     urlfieldinverter.cpp
     wordstore.cpp
     DEPENDS
