+++ searchlib/src/vespa/searchlib/memoryindex/posting_iterator.h
